# docs/swagger.yaml
basePath: /
definitions:
  models.BadRequestResponse:
    description: Standard error response object
    properties:
      bad request:
        type: string
    type: object
  models.CreateAPIEventRequestSwagger:
    description: CreateAPIEventRequestSwagger used only for Swagger docs
    properties:
      api_endpoint:
        example: https://api.example.com/v1/users
        type: string
      api_method:
        enum:
        - GET
        - POST
        - PUT
        - PATCH
        - DELETE
        example: GET
        type: string
      api_payload:
        additionalProperties:
          type: string
        type: object
      name:
        example: API Event 1
        type: string
      type:
        enum:
        - API
        example: API
        type: string
    required:
    - api_endpoint
    - api_method
    - name
    - type
    type: object
  models.CreateUserRequest:
    description: CreateUserRequest object used for input
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - full_name
    - password
    - username
    type: object
  models.InternalServerErrorResponse:
    description: Standard error response object
    properties:
      internal server error:
        type: string
    type: object
  models.LoginRequest:
    description: LoginRequest object used for input
    properties:
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.LoginResponse:
    description: LoginResponse object used for output
    properties:
      access_token:
        type: string
      access_token_expiry:
        type: string
      refresh_token:
        type: string
      refresh_token_expiry:
        type: string
      session_id:
        type: string
      user:
        $ref: '#/definitions/models.UserDto'
    type: object
  models.NotFoundResponse:
    description: Standard error response object
    properties:
      not found:
        type: string
    type: object
  models.RenewAccessTokenRequest:
    description: RenewAccessTokenRequest object used for input
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  models.RenewAccessTokenResponse:
    description: RenewAccessTokenResponse object used for input
    properties:
      access_token:
        type: string
      access_token_expiry:
        type: string
    type: object
  models.SwaggerEventDto:
    description: SwaggerEventDto used only for Swagger docs
    properties:
      after_x_mins:
        type: integer
      api_endpoint:
        type: string
      api_method:
        type: string
      api_request_body:
        additionalProperties:
          type: string
        type: object
      created_by:
        type: string
      createdAt:
        type: string
      executed_at:
        type: string
      id:
        type: integer
      interval:
        type: integer
      is_recurring:
        type: boolean
      name:
        type: string
      run_at:
        type: string
      type:
        type: string
    type: object
  models.UnauthorizedRequestResponse:
    description: Standard error response object
    properties:
      unauthorized request:
        type: string
    type: object
  models.UserDto:
    description: UserDto object used for input
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      password_changed_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: Musthafa
    url: http://www.swagger.io/support
  description: This is an event scheduling service with JWT authentication
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Event Scheduler API
  version: "1.0"
paths:
  /events:
    get:
      description: List all events
      parameters:
      - description: Page Number
        in: query
        minimum: 1
        name: pageNumber
        required: true
        type: integer
      - description: Page Size
        in: query
        minimum: 1
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            description: SwaggerListEventResponse used only for Swagger docs
            items:
              $ref: '#/definitions/models.SwaggerEventDto'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.UnauthorizedRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: List Events
      tags:
      - Events
  /events/api:
    post:
      consumes:
      - application/json
      description: Create a new API Event
      parameters:
      - description: API Event data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAPIEventRequestSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwaggerEventDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.UnauthorizedRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: Create API Event
      tags:
      - Events
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate and get access/refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.UnauthorizedRequestResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      summary: Login user
      tags:
      - Auth
  /token/renew:
    post:
      consumes:
      - application/json
      description: Get a new access token using the refresh token
      parameters:
      - description: Token data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RenewAccessTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RenewAccessTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.UnauthorizedRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      summary: Renew Access Token
      tags:
      - Auth
  /users:
    post:
      consumes:
      - application/json
      description: Create a new user with username, password, and email
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      summary: Create a new user
      tags:
      - Auth
  /users/{username}:
    delete:
      consumes:
      - application/json
      description: Remove a user by username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.UnauthorizedRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user by username
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Retrieve user details by username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.UnauthorizedRequestResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by username
      tags:
      - Users
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'JWT token authorization format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"